remove `calc data`
